cap nhat entity course

+++ BE/KLTN_final/KLTN_final/src/main/java/com/vlearning/KLTN_final/domain/dto/response/Instructor.java
@@ -13,5 +13,5 @@ public class Instructor extends User {
 
     private Integer totalCourses;
 
-    private Integer totalRating;
+    private Float totalRating;
 }
